requirements.txt
requests
pandas
boto3
minio
pymongo
matplotlib
plotly
python-dotenv
xenaPython
